Move gpu cache logic into a module
